refactor: renames momento-vector-index to momento_vector_index
Per style conventions, prefer snake case for modules over kebab
case. We update the module, tests, examples, docs, and name exports
per this convention.

We re-ran integration tests and examples to confirm working. We also
looked at the docs pages to confirm correct.
