+++ langchain/.gitignore
@@ -184,9 +184,9 @@ vectorstores/weaviate.d.ts
 vectorstores/lancedb.cjs
 vectorstores/lancedb.js
 vectorstores/lancedb.d.ts
-vectorstores/momento-vector-index.cjs
-vectorstores/momento-vector-index.js
-vectorstores/momento-vector-index.d.ts
+vectorstores/momento_vector_index.cjs
+vectorstores/momento_vector_index.js
+vectorstores/momento_vector_index.d.ts
 vectorstores/mongo.cjs
 vectorstores/mongo.js
 vectorstores/mongo.d.ts
