+++ langchain/scripts/create-entrypoints.js
@@ -78,7 +78,7 @@ const entrypoints = {
   "vectorstores/faiss": "vectorstores/faiss",
   "vectorstores/weaviate": "vectorstores/weaviate",
   "vectorstores/lancedb": "vectorstores/lancedb",
-  "vectorstores/momento-vector-index": "vectorstores/momento-vector-index",
+  "vectorstores/momento_vector_index": "vectorstores/momento_vector_index",
   "vectorstores/mongo": "vectorstores/mongo",
   "vectorstores/mongodb_atlas": "vectorstores/mongodb_atlas",
   "vectorstores/pinecone": "vectorstores/pinecone",
@@ -307,7 +307,7 @@ const requiresOptionalDependency = [
   "vectorstores/weaviate",
   "vectorstores/lancedb",
   "vectorstores/mongo",
-  "vectorstores/momento-vector-index",
+  "vectorstores/momento_vector_index",
   "vectorstores/mongodb_atlas",
   "vectorstores/pinecone",
   "vectorstores/qdrant",
